Added Alternate Navbar
The navbar for Demo and help pages is not under the control of Devise,
so an alternate navbar is necessary to avoid the page deciding that
logging out on these pages is a forgery. To this end, the logout button
on the alternate navbar has been replaced with a "Return to application"
button, which just takes the user to the home page.