Demo Navbar & new tasks
We can now create tasks, once engine makes the appropriate changes to
the controller on their end. This task will be linked to the proper
hunt.

The navbar has also been changed. On the demo page, you cannot log out,
as this will throw an exception with Devise. So, that button changed.