Changed button on Create Hunt page
Changed the size of the buttons overall (I think), "fixed" the @media
stuff now that I understand how it actually works. Can't get the
'submit' type buttons centered in the form, but that's minor. Not sure
how much size scaling we'll actually need on the buttons because the
screen sizes for most phones aren't too terribly different. It does
change depending on phone and large-screen tablet/computer.