Bug Fixes and Footer
The footer now displays at the bottom of every page.
Additionally, the back to top button now returns to wherever the header
was rendered, which should be the top of the page.

Also, minor bug fixes:
Hunt new form now displays an end at tag
closed an open div in the hunts show page