Fixed lolcat validation bug
When you changed task types in the middle of validation, the form filled
in the correct answer field with "lolcat", the filler value we had used
for validation for the correct answer field. This no longer happens, as
we no longer need a filler value because the validation is now working
more efficiently.